Completed WBS 528Row (#780)

+++ src/views/components/admin/group/organisms/AdminDeleteGroup.tsx
@@ -10,7 +10,7 @@ import ButtonNegative from "views/components/common/molecules/ButtonNegative";
 
 type Props = {
   companyId: string;
-  groupLocalId: string;
+  groupId: string;
   id: string;
 };
 
@@ -23,7 +23,7 @@ export default function (props: Props) {
           companyID: props.companyId,
           filter: {
             groupID: {
-              eq: props.groupLocalId,
+              eq: props.groupId,
             },
           },
         };
@@ -32,23 +32,29 @@ export default function (props: Props) {
           listI
         );
 
-        if (employeeItem && employeeItem.length === 0) {
-          const deleteI: DeleteGroupInput = {
-            id: props.id,
-          };
-          const deleteItem = await GroupDao.deleteByAdmin(
-            deleteGroupByCompanyAdmin,
-            deleteI
-          );
-
-          if (deleteItem) {
-            window.alert("部署の削除が完了しました");
-            window.location.href = routeBuilder.adminGroupListPath();
-          } else {
-            setError("部署情報の削除に失敗しました");
+        if (window.confirm("部署情報を削除してもよろしいでしょうか？")) {
+          if (!employeeItem) {
+            setError("社員情報の取得に失敗しました");
+            return;
           }
-        } else {
-          window.alert("部署に社員が存在するため部署を削除できません");
+          if (employeeItem && employeeItem.length === 0) {
+            const deleteI: DeleteGroupInput = {
+              id: props.id,
+            };
+            const deleteItem = await GroupDao.deleteByAdmin(
+              deleteGroupByCompanyAdmin,
+              deleteI
+            );
+
+            if (deleteItem) {
+              window.alert("部署の削除が完了しました");
+              window.location.href = routeBuilder.adminGroupListPath();
+            } else {
+              setError("部署情報の削除に失敗しました");
+            }
+          } else {
+            window.alert("部署に社員が存在するため部署を削除できません");
+          }
         }
       }}
     >
